+++ packages/common/src/graphql/entities.ts
@@ -60,7 +60,7 @@ export function getAllEntitiesRelations(_schema: GraphQLSchema | string): GraphQ
     for (const field of Object.values(entity.getFields())) {
       const typeString = extractType(field.type).toString();
       const derivedFromDirectValues = getDirectiveValues(derivedFrom, field.astNode);
-
+      const indexDirectiveVal = getDirectiveValues(indexDirective, field.astNode);
       //If is a basic scalar type
       if (Object.values(FieldScalar).includes(typeString)) {
         newModel.fields.push(packEntityField(typeString, field, false));
@@ -74,6 +74,13 @@ export function getAllEntitiesRelations(_schema: GraphQLSchema | string): GraphQ
           to: typeString,
           foreignKey: `${field.name}Id`,
         } as GraphQLRelationsType);
+        if (typeString !== 'ID') {
+          newModel.indexes.push({
+            unique: false,
+            fields: [`${field.name}Id`],
+            using: IndexType.HASH,
+          });
+        }
       }
       // If is derivedFrom
       else if (entityNameSet.includes(typeString) && derivedFromDirectValues) {
@@ -101,7 +108,6 @@ export function getAllEntitiesRelations(_schema: GraphQLSchema | string): GraphQ
         throw new Error(`${typeString} is not an valid type`);
       }
       // handle indexes
-      const indexDirectiveVal = getDirectiveValues(indexDirective, field.astNode);
       if (indexDirectiveVal) {
         if (typeString !== 'ID' && Object.values(FieldScalar).includes(typeString)) {
           newModel.indexes.push({
@@ -109,11 +115,8 @@ export function getAllEntitiesRelations(_schema: GraphQLSchema | string): GraphQ
             fields: [field.name],
           });
         } else if (typeString !== 'ID' && entityNameSet.includes(typeString)) {
-          newModel.indexes.push({
-            unique: indexDirectiveVal.unique,
-            fields: [`${field.name}Id`],
-            using: IndexType.HASH,
-          });
+          //is foreign key
+          continue;
         } else {
           throw new Error(`index can not be added on field ${field.name}`);
         }